Make DpadAxisControl an internal nested class of DpadControl
