+++ Packages/com.unity.inputsystem/InputSystem/Controls/DpadControl.cs
@@ -17,6 +17,28 @@ namespace UnityEngine.Experimental.Input.Controls
     /// </remarks>
     public class DpadControl : Vector2Control
     {
+        internal class DpadAxisControl : AxisControl
+        {
+            public int component;
+
+            protected override void FinishSetup(InputDeviceBuilder builder)
+            {
+                base.FinishSetup(builder);
+                component = name == "x" ? 0 : 1;
+
+                // Set the state block to be the parent's state block. We don't use that to read
+                // the axis directly (we call the parent control to do that), but we need to set
+                // it up the actions know to monitor this memory for changes to the control.
+                m_StateBlock = m_Parent.m_StateBlock;
+            }
+
+            public override unsafe float ReadUnprocessedValueFromState(void* statePtr)
+            {
+                var value = (m_Parent as DpadControl).ReadUnprocessedValueFromState(statePtr);
+                return value[component];
+            }
+        }
+
         // The DpadAxisControl has it's own logic to read state from the parent dpad.
         // The useStateFrom argument here is not actually used by that. The only reason
         // it is set up here is to avoid any state bytes being reserved for the DpadAxisControl.
